Refactoring, to separate client handlers
